refactor: split interrupt.c (794 lines) into 3 focused modules
- interrupt.c: core IDT setup + dispatch (566 lines)
- interrupt_exceptions.c: exception handlers — double fault, NMI, page fault,
  GPF + panic dump (152 lines)
- interrupt_syscall.c: syscall registration + management table (77 lines)

Makefile reorganized into 6 modular includes under mk/;
fixes duplicate test_mime target (was test_trash at L851)

+++ src/runtime/wubu_6502_interp.c
@@ -297,5 +297,7 @@ int64_t wubu_6502_run(const uint8_t *code, size_t size, int64_t arg)
             break;
         }
     }
-    return (int64_t)cpu.a;
+    /* sign-extend the 8-bit accumulator so the battery's negative
+     * expectations (-1 for ~0, etc.) agree with the other drivers */
+    return (int64_t)(int8_t)cpu.a;
 }
